app/core/config.py
#!/usr/bin/env python
# -*- coding:utf-8 -*-
# @Author:      dudea
# @Time:        2022/5/9 11:14
import os

from dotenv import load_dotenv
from pydantic import BaseSettings

load_dotenv()


class Settings(BaseSettings):
    API_V1_STR: str = "/api/v1"
    PROJECT_NAME: str = "cndbase"
    BASICEG_URI: str = os.getenv("ZYSL_BASICEG_HTTP")


settings = Settings()
