requirements.txt
anyio==3.6.1
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
ecdsa==0.17.0
fastapi==0.78.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
install==1.3.5
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.18.1
